refactor building a cryptarithm node from a gen. model #32
